# Optimize set of nodeID for basic scan node table queries