--- a/src/storage/store/node_table.cpp
+++ b/src/storage/store/node_table.cpp
@@ -207,8 +207,9 @@ bool NodeTableScanState::scanNext(Transaction* transaction) {
         nodeGroupStartOffset = transaction->getUncommittedOffset(tableID, nodeGroupStartOffset);
     }
     for (auto i = 0u; i < scanResult.numRows; i++) {
-        nodeIDVector->setValue(i,
-            nodeID_t{nodeGroupStartOffset + scanResult.startRow + i, tableID});
+        auto& nodeID = nodeIDVector->getValue<nodeID_t>(i);
+        nodeID.tableID = tableID;
+        nodeID.offset = nodeGroupStartOffset + scanResult.startRow + i;
     }
     return true;
 }
@@ -511,7 +512,7 @@ void NodeTable::commit(Transaction* transaction, TableCatalogEntry* tableEntry,
     // 2. Set deleted flag for tuples that are deleted in local storage.
     row_idx_t numLocalRows = 0u;
     for (auto localNodeGroupIdx = 0u; localNodeGroupIdx < localNodeTable.getNumNodeGroups();
-         localNodeGroupIdx++) {
+        localNodeGroupIdx++) {
         const auto localNodeGroup = localNodeTable.getNodeGroup(localNodeGroupIdx);
         if (localNodeGroup->hasDeletions(transaction)) {
             // TODO(Guodong): Assume local storage is small here. Should optimize the loop away by
@@ -651,7 +652,7 @@ void NodeTable::scanPKColumn(const Transaction* transaction, PKColumnScanHelper&
 
     const auto numNodeGroups = nodeGroups_.getNumNodeGroups();
     for (node_group_idx_t nodeGroupToScan = 0u; nodeGroupToScan < numNodeGroups;
-         ++nodeGroupToScan) {
+        ++nodeGroupToScan) {
         scanState->nodeGroup = nodeGroups_.getNodeGroupNoLock(nodeGroupToScan);
 
         // It is possible for the node group to have no chunked groups if we are rolling back due to