# Description

The large num of constructing and copying of nodeID_t objects impose quite large overheads for basic node table scans.

On my local Mac mini desktop, on ldbc-100, the following query can be improved quite a bit.
```
MATCH (comment:Comment) WHERE comment.length < 150 RETURN count(*);
```

| master | this branch |
| ------- | ----------- |
| 68.84ms | 37.13ms  |

TODO:
- There could be more places we're allocating and copying internalID_t objects. We should rewrite them all.